llmstxtgenerator/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pyyaml==6.0.1
pandas==2.1.0
tqdm==4.66.1
urllib3==2.0.4
validators==0.22.0
tiktoken==0.5.1
openai==1.6.1
anthropic==0.8.1